+++ addon/collections/promise.js
@@ -1,0 +1,2 @@
+import Ember from 'ember';
+export default Ember.Object.extend(Ember.PromiseProxyMixin);