+++ tests/unit/collections/indexed-collection-test.js
@@ -102,19 +102,22 @@ module("IndexedCollection - fetching", {
   }
 });
 
-test("fetch returns a promise which resolves when the collection has a value", function() {
-  expect(2);
+test("fetch returns a promise proxy which resolves when the collection has a value", function() {
+  expect(4);
 
   var people = PeopleIndex.create();
+  var promise = people.fetch();
 
-  people.fetch().then(function(c) {
+  promise.then(function(c) {
     equal(c, people, "resolves with itself");
     equal(people.get("length"), 3, "has the items");
   });
 
   // we need two flushes because the first kicks off the fetches for the objects
+  ok(promise.get("isPending"));
   firebase.flush();
   firebase.flush();
+  ok(!promise.get("isPending"));
 });
 
 test("resolves immediately if already listening to firebase", function() {